ci: Add google-chrome image build to CI workflow
This commit adds a step to the CI workflow to build a Docker image
for google-chrome. The new job is configured in the CI file
.github/workflows/containers.yml and utilizes the `aardsoft/ci-actions/kiwi-build`
action.

The configuration is defined in the newly created file
tw-wayland/google-chrome/appliance.kiwi. This XML specifies the
parameters for creating a SUSE-derived Docker container
including the repository and package details.

The package manager is set to `zypper`, and `google-chrome-stable` is
included in the image packages. This change aims to automate the
containerization of google-chrome, easing deployment and testing processes.